chore: rename lib to redux
