+++ app/components/(functional)/MobileMenu/MobileMenu.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import { useAppDispatch } from "@/lib/hooks";
+import { useAppDispatch } from "@/app/redux/hooks";
 import { MobileAccordion } from "@/app/components";
 import { routes, authRoutes } from "@/app/constants";
 import { MobileMenuProps } from "@/app/types";
-import { clearSession, setAuthenticated } from "@/lib/features/Auth";
+import { clearSession, setAuthenticated } from "@/app/redux/features/Auth";
 import styles from "./MobileMenu.module.css";
 
 export default function MobileMenu({ toggleMenu, role }: MobileMenuProps) {
